refactor: move operator into operators module and remove legacy entry
