--- PlanetoPBR.py
@@ -1,44 +1,0 @@
-from .utils import (
-    load_pbr_textures,
-    import_plane_from_image,
-)
-
-bl_info = {
-    "name": "planetopbr",
-    "blender": (2, 80, 0),
-    "category": "Object",
-}
-
-# Going rem,ove keep for now
-import bpy
-import os
-# Going to remove keep for now
-from bpy_extras.io_utils import ImportHelper
-
-
-class OBJECT_OT_import_plane_from_image(bpy.types.Operator, ImportHelper):
-    """Operator to import a plane with a PBR material using an image as reference."""
-    bl_idname = "object.import_plane_from_image"
-    bl_label = "Import Plane from Image"
-    bl_options = {'REGISTER', 'UNDO'}
-
-    directory: bpy.props.StringProperty(subtype="DIR_PATH")
-
-    def execute(self, context):
-        """Execute the operator and import the plane from the diffuse image."""
-        textures = load_pbr_textures(self.directory)
-        import_plane_from_image(textures)
-        return {'FINISHED'}
-
-    def invoke(self, context, event):
-        """Invoke the directory selection dialog."""
-        context.window_manager.fileselect_add(self)
-        return {'RUNNING_MODAL'}
-
-def menu_func(self, context):
-    """Add the operator to the 'Add Mesh' menu."""
-    self.layout.operator(OBJECT_OT_import_plane_from_image.bl_idname)
-
-
-# if __name__ == "__main__":
-#     register()
